Refactor test helpers to accomidate multiple versions
Adds an interface for v1 and v2 of the contracts state and contracts, to be updated as more functionality is added in.
Each test then just loads the appropriate set of tests and state from the fixture. e.g.

    contracts = testSetup.contracts
    state = testSetup.state

becomes

    contracts = testSetup.contracts.v1
    state = testSetup.state.v1

Deployment scripts can be moved into
versioned directories. Slither complained about both contracts having the same name

https://github.com/crytic/slither/wiki/Detector-Documentation#name-reused

So renamed TroveManager to TroveManagerV2.