Update protocol loan repayment logic
Renamed payDebt to distributeMUSD.

if no feeRecipient is set and there is a loan then 100% of the amount
goes to paying the loan.

if no feeRecipient is set and the loan is paid 100% is deposited into
the stabilityPool.

if a feeRecipient is set and there is a loan then part is used to repay
the loan (min 50%) and the rest is sent to the feeRecipient.

if a feeRecipient is set and there is no loan then part is deposited
into the stabilityPool and the reset is sent to the feeRecipient.

When there is a loan and the final repayment is made which is greater
than the loan the excess is deposited into the StabilityPool.