Separate initialization from constructor for MUSD token
Introduced `initialize` function setting system contracts outside of the
constructor. This will allow the deployer contract to establish stable
address for the token contract, irrespective of the system contract
addresses.

`initialize` can be called only one time by the owner, which at the time
of the deployment will be the deployer.