Explain where the deployer and governance addresses come from
Those addresses are hardcoded in the TokenDeployer contract to ensure
the same bytecode is used for stable deployments across all networks.

The deployer EOA is the same as the one used to deploy tBTC v1, tBTC v2,
and all Mezo contracts.

The governance multisig is the same as the one used to control the
upgradeability of Mezo contracts as well as some of the chain
parameters.