Let deployer EOA to deploy token only on Mezo and Ethereum
On all other chains, the governance multisig is required.

There were concerns regarding using the deployer address long-term.
If, after some time, we wanted to deploy token to another chain and were
unlucky to have the deployer account compromised, we wouldn't be able to
achieve stable addresses across the chains. What is worse, the attacker
would be able deploy the contract to another chain with malicious
initializer parameters under the same address we had initially on Mezo.
The governance would be able to alter it but that would require
deploying the governance, 24h delay, and first of all, knowing
a malicious token instance got deployed.

The governance multisig is not an ideal option either. First, it slows
down the deployment time pretty significantly. For Mezo specifically,
the Safe Factory contracts with three levels of dependencies have to be
deployed. Then we create the governance Safe by mirroring the past
transaction. Then, we need to assemble the transaction to deploy the
token, confirm, and sign it, which, without Safe UI support, is a really
painful process. We have to do the first step regardless, but we can
do it later with EOA. Second, the past Safe creation transaction we are
mirroring contains the initial set of owners, and only with those initial
owners the Safe can be deployed under the same address. If the original
signers are lost, we can't issue transactions from the Safe on the new
chain.

This commit is a middle ground. We let the existing deployer to deploy
the mUSD contract on Mezo and Ethereum by checking the chain ID. This is
something we want to do right away. For all other chains, we will
require the governance to deploy.