Remove governance delay from the MUSD token contract
The governance delay was used to control the burn list and the mint
list. Given that we want to deploy the token with a shorter governance
delay, there was a need to increase that delay later and implement the
functionality to achieve it, keeping in mind that the governance delay
change should be guarded by the current governance delay. In addition
to that, we had concerns about whether this functionality should exist
in the token or the newly introduced GovernableVariables contract.

All of this is quite complicated and error-prone. It is especially
alarming given the tooling for enforcing governance delays exists and is
widely used in the space in the form of timelock controllers.

This change removes the governance delay from the token altogether.
Initially, the token contract will be controlled by 5/9 governance, just
like the chain. Eventually, once we gain more confidence in the system,
we will deploy a timelock controller. This functionality is not provided
in this changeset and can be deployed later, at any point in time,
separately.

I took the chance and applied some additional fixes:
- We could add to the mint list addresses that were not system
  contracts, but the same was not possible for the burn list. Now both
  options are available, which makes the token contract more
  future-proof.
- No events for adding to the burn and mint list were emitted. Added
  them.
- Removed unused event from IMUSD and cleared up the interface. The
  initialization and governance functions should not be in the area of
  interest of integrations.