Simplified upgradeability tests for PriceFeed
For upgradeability tests, we can trust the toolkit and the virtual
machine. The address is not going to change, and the contract will be
usable from other contracts. What we need is a very basic test ensuring
`helpers.upgrades.upgradeProxy` does not fail.

Also, `PriceFeedUpgradeTester` was not entirely correct :) There is no
need for a gap in that contract, and `initialize(address _owner)` was
not called. If it was called, it would fail as the version 1 was
already initialized (the `initializer` modifier).