Basic upgradeability tests for BorrowerOperationsSignatures
Those tests are a bit more complicated than the rest of the tests in the
suite, given that we want to check signatures prepared for the old
implementation are no longer accepted, and signatures prepared for the
new implementation work as expected.

I also took the opportunity and improved all other tests a bit by
extracting common upgrade logic to one function and adding a sanity
check to each tests that the address did not change.