Remove the guard for `msg.sender == this` in `adjustTrove`
This is a vestigage of how we were previously handling the `restricted`
functions. Previously, they were `public` functions that could only be called
by:

- BorrowerOperationsSignatures
- BorrowerOpeartions
- The stability pool

We refactored this into external functions that users can call, and external
functions that only BorrowerOperationsSignatures can call, and both of those
functions make a call to an internal function to reduce duplication. The
internal functions no longer need to check to see if the caller is
BorrowerOperations, because it always is.