Rip out unused `totalCollSurplus` from the liquidation flow
Previously, recovery mode liquidations could be performed on troves under 150%
ICR (instead of under 110%), but would only liquidate enough collateral to
cover 110% of the debt. The remaining collateral was sent to the
CollSurplusPool, where it waited for its owner to pick it up.

Now, liquidations in recovery mode follow the same rules as normal mode, so
there is no excess collateral. Thus, we can safely remove this code path.