RFC-1: Bridging MUSD to Ethereum
This RFC aims at providing the infrastructure required to bridge MUSD minted on
Mezo to Ethereum, enabling third-party integrations to implement various yield
strategies.

The MUSD token is deployed on Mezo and Ethereum under the same address
`0xdD468A1DDc392dcdbEf6db6e34E89AA338F9F186`. We are going to use Wormhole for
bridging, but instead of ending the bridging flow with a Wormhole-wrapped token
as MUSD on Ethereum, we are going to mint the canonical MUSD. This mechanism is
implemented and battle-proven for tBTC bridging from Ethereum to various L2s.