RFC-2: Lending Arbitrage Microloans
MUSD's minimum debt size of 1800 excludes many users. Some may want to borrow
smaller amounts to try out the system before taking out a larger loan, or they
simply may not need to borrow that much MUSD.

We are seeking to increase the overall MUSD volume by allowing users access to
MUSD loans with a much smaller minimum (e.g. $25).

As a secondary goal, keeping the user experience (at least the frontend) close
to the MUSD borrow experience is valuable so that users can get a feeling for
how the system works if they do choose to borrow more later. It is not critical
that the systems be exactly 1 to 1, but if we can present them similarly, that
is a plus.

We generally expect to see two types of users borrowing the smaller amounts.
The first type of user will borrow a small amount of MUSD with the minimum
possible collateralization and spend that MUSD. They will not come back to
manage the collateral or repay the debt. The second type of user will mint
a small amount of MUSD with a relatively safe collateralization and play with
the system. They will keep an eye on collateralization, and after some time,
when they are comfortable using the system, they will come back to put more
money in. 

The RFC-1: Microloans was the first attempt at solving this problem by
introducing a separate smart contract allowing to borrow any amount below the
current's MUSD protocol minimum of 1800 and above the minimum proposed microloan
size of 50 MUSD. This flexibility came from the fact, the smart contract offering
microloans had its own trove open in the MUSD protocol and could mint any amount
of MUSD based on the provided collateral.

That flexibility came with the cost. After gaining traction and enough users
trying the system with the minimum collateralization ratio allowed, the trove
controlled by the microloans contract is at constant risk of being a target of
external MUSD redemptions. With the collateral removed from the trove,
protecting the user experience - especially for the second user type who wants
to come back and repay their debt - is complicated. 

RFC-2: Lending Arbitrage Microloans is an attempt to offer microloans in a way
that is protecting the user experience at the cost of scalability of the first
release. Possible solutions for scaling, after getting enough traction, are
proposed as the next steps but they are out of the scope of this RFC.