Modifying withdrawMUSD function without accounting for MUSD protocol debt
Protocol debt repayment will be handled differently, not paid through
MUSD protocol fees. But we also need a function to transfer existing
collected fees to some escrow contract for further distribution as
transferring all the existing fees to MUSD savings vault creates a
situation when all the fees can be taken by a quick deposit and
withdraw of a small MUSD amount. How are we going to distribute existing
fees is yet to be decided.