Keep all PCV fund flow functions next to each other
It is easier to reason about the flow of funds if all those functions
are place next to each other. No logic changes in this commit, just
reordering functions to keep all set/add etc governance functions at the
top and have all the function controlling the flow of funds next.