Send tokens to recipient when PCV withdraws from StabilityPool
Previously we were withdrawing to the PCV contract but it conflicts with
the distribute* functions. Since distribute* functions will eventually
be permissionless, we would always need to withdraw collateral and MUSD
in the same transaction as withdrawFromStabilityPool, to not let those
funds be distributed as yield to Tigris.

This change adds a recipient parameter to withdrawFromStabilityPool to
receive the funds. Note it does not change anything security-wise, as we
still require the recipient to be whitelisted.