Deposit to StabilityPool via PCV with external tokens
The previous implementation of depositToStabilityPool assumed MUSD
deposited to StabilityPool comes from the PCV contract balance. It does
not make a lot of sense in the new setup, as MUSD balance on PCV
is distributed by distributeMUSD. So instead of worrying about funding
PCV and calling depositToStabilityPool in the same TX, let's just take
funds approved by msg.sender to deposit.