Removed PCV withdrawBTC and withdrawBTC functions
The PCV contract no longer accumulates funds. The yield is
distributed to feeRecipient and btcRecipient. Whatever stays in the
contract, as set in feeSplitPercentage, is used to either repay the
bootstrap loan or deposited to the stability pool. When collateral or
MUSD are withdrawn from the stability pool, they go straight to the
recipient passed as a parameter to withdrawFromStabilityPool. Since
nothing that should not be distributed accumulates in PCV, the functions
can be removed.

Note that if governance ever decides to use fees to
increase the stability pool position before the debt is paid, they
can still do that by setting feeRecipient to an escrow contract.